// ast-schema/generated/serialize.ts
export interface JSONNode {
  kind: string;
  pos: number;
  end: number;
  text?: string;
  children?: JSONNode[];
  leadingComments?: KSCommentRange[];
  trailingComments?: KSCommentRange[];
  [key: string]: unknown;
}